# backend/openapi/openapi.internal.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/syncdrive-api/operation-shift/list": {
      "get": {
        "operationId": "OperationShiftController_listShifts",
        "parameters": [
          {
            "name": "keyword",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "usage_status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "publish_status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page_size",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "正線班表列表（分頁、篩選）",
        "tags": [
          "Operation Shifts"
        ]
      }
    },
    "/syncdrive-api/operation-shift/check-name": {
      "get": {
        "operationId": "OperationShiftController_checkName",
        "parameters": [
          {
            "name": "name",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "exclude_id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "檢查班表名稱是否唯一",
        "tags": [
          "Operation Shifts"
        ]
      }
    },
    "/syncdrive-api/operation-shift/timetable/trips": {
      "get": {
        "description": "回傳目前使用中班表之全部班次，含各班次的逐站計畫抵達與發車時刻。任務類型涵蓋正線、保養、行檢、充電、待命與調度。本端點為計畫值，與車輛實際位置無關；即時值見車輛即時 ETA。",
        "operationId": "OperationShiftController_getTimetableTrips",
        "parameters": [
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "HH:MM:SS 或秒，預設 00:00:00",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "HH:MM:SS 或秒，預設 24:00:00",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimetableTripsResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "缺少或不正確的 x-api-key 標頭"
          },
          "404": {
            "description": "庫內沒有任何含排班結果的班表"
          }
        },
        "security": [
          {
            "x-api-key": []
          }
        ],
        "summary": "取得班次清單與各站計畫時刻",
        "tags": [
          "Operation Shifts",
          "對外｜班表計畫"
        ]
      }
    },
    "/syncdrive-api/operation-shift/timetable/station-etas": {
      "get": {
        "description": "以停靠點為主鍵，回傳由班表推算之抵達與離站事件序列。本端點為計畫值，事件不帶車輛識別；需要車輛識別時使用車輛即時 ETA。可用 from／to 限定時間區間，station_id 限定單一停靠點。",
        "operationId": "OperationShiftController_getStationEtas",
        "parameters": [
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "HH:MM:SS 或秒，預設 00:00:00",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "HH:MM:SS 或秒，預設 24:00:00",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "station_id",
            "required": false,
            "in": "query",
            "description": "只取單一站",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StationEtasResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "缺少或不正確的 x-api-key 標頭"
          },
          "404": {
            "description": "庫內沒有任何含排班結果的班表"
          }
        },
        "security": [
          {
            "x-api-key": []
          }
        ],
        "summary": "取得各停靠點之計畫事件流",
        "tags": [
          "Operation Shifts",
          "對外｜班表計畫"
        ]
      }
    },
    "/syncdrive-api/operation-shift/detail/{id}": {
      "get": {
        "operationId": "OperationShiftController_getShiftDetail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "取得班表詳情（含 body）",
        "tags": [
          "Operation Shifts"
        ]
      },
      "patch": {
        "operationId": "OperationShiftController_updateDraft",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "更新班表草稿",
        "tags": [
          "Operation Shifts"
        ]
      },
      "delete": {
        "operationId": "OperationShiftController_deleteShift",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "刪除班表",
        "tags": [
          "Operation Shifts"
        ]
      }
    },
...
    "/syncdrive-api/operation-shift/detail/{id}/publish": {
      "post": {
        "operationId": "OperationShiftController_publishShift",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "發布班表（供 timetable API 優先讀取）",
        "tags": [
          "Operation Shifts"
        ]
      }
    },
    "/syncdrive-api/operation-shift/detail/{id}/deploy": {
      "post": {
        "operationId": "OperationShiftController_deployShift",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "部署班表為使用中（同時發布；其他使用中班表改為閒置）",
        "tags": [
          "Operation Shifts"
        ]
      }
    },
    "/syncdrive-api/operation-shift/detail/{id}/duplicate": {
      "post": {
        "operationId": "OperationShiftController_duplicateDraft",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "以此複製新版班表草稿",
        "tags": [
          "Operation Shifts"
        ]
      }
    },
    "/syncdrive-api/operation-shift/detail/{id}/duplicate-as-manual": {
      "post": {
        "operationId": "OperationShiftController_duplicateAsManualDraft",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "參數生成班表複製成手動製作草稿",
        "tags": [
          "Operation Shifts"
        ]
      }
    },
...
  },
  "info": {
    "title": "SyncDrive-T3 API（內部）",
    "description": "我方前端、後台與維運使用的完整介面。**這一份不對外發布**——交付給協力廠商的是對外文件（見 /api/docs/public）。",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "x-api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
    "schemas": {
...
      "EtaPlanDto": {
        "type": "object",
        "properties": {
          "planned_arrival_at": {
            "type": "object",
            "nullable": true,
            "example": 1786842010000
          },
          "planned_arrival_clock": {
            "type": "object",
            "nullable": true,
            "example": "09:00:10"
          },
          "planned_departure_at": {
            "type": "object",
            "nullable": true,
            "description": "僅於該停靠點為此班次起站，或 arrival_state 為 AT_STATION 時有值"
          },
          "planned_departure_clock": {
            "type": "object",
            "nullable": true
          },
          "delay_seconds": {
            "type": "object",
            "nullable": true,
            "example": 13,
            "description": "eta_at 減 planned_arrival_at，單位秒。正值＝晚於計畫"
          },
          "delay_state": {
            "type": "string",
            "enum": [
              "EARLY",
              "ON_TIME",
              "MINOR_DELAY",
              "MAJOR_DELAY",
              "NO_PLAN"
            ],
            "description": "暫定定義，門檻未經營運確認。現階段請以 delay_seconds 為準"
          }
        },
        "required": [
          "planned_arrival_at",
          "planned_arrival_clock",
          "planned_departure_at",
          "planned_departure_clock",
          "delay_seconds",
          "delay_state"
        ]
      },
      "StationEtaEntryDto": {
        "type": "object",
        "properties": {
          "vehicle_code": {
            "type": "string",
            "example": "PMS-05"
          },
          "order_id": {
            "type": "object",
            "nullable": true,
            "example": "260816-ST0007"
          },
          "trip_code": {
            "type": "object",
            "nullable": true,
            "example": "ST0007",
            "description": "路線代號＋發車時刻 HHMM"
          },
          "route_code": {
            "type": "object",
            "nullable": true,
            "example": "ST"
          },
          "route_name": {
            "type": "object",
            "nullable": true,
            "example": "S2W上行 > T3上行"
          },
          "vehicle_phase": {
            "type": "string",
            "nullable": true,
            "enum": [
              "TRANSITING",
              "FAULTED"
            ],
            "description": "車端營運階段，原值透傳；資料逾時或車端未提供時為 null"
          },
          "arrival_state": {
            "type": "string",
            "enum": [
              "EN_ROUTE",
              "APPROACHING",
              "DOCKING",
              "AT_STATION",
              "DEPARTED",
              "UNKNOWN"
            ],
            "description": "APPROACHING＝即將進站（eta ≤ 60 秒或距離 ≤ 200 公尺）；EN_ROUTE＝預計到達；UNKNOWN＝車端資料逾時或缺少目標站"
          },
          "eta_seconds": {
            "type": "object",
            "nullable": true,
            "example": 25,
            "description": "以 observed_at 為基準的剩餘秒數，非以收到回應的時刻為基準"
          },
          "eta_at": {
            "type": "object",
            "nullable": true,
            "example": 1786842023000,
            "description": "恆等於 observed_at + eta_seconds × 1000。倒數計時必須用這個欄位"
          },
          "eta_clock": {
            "type": "object",
            "nullable": true,
            "example": "09:00:23"
          },
          "distance_to_station_m": {
            "type": "object",
            "nullable": true,
            "example": 120,
            "description": "路徑距離，公尺"
          },
          "plan": {
            "$ref": "#/components/schemas/EtaPlanDto"
          },
          "observed_at": {
            "type": "number",
            "example": 1786841998000,
            "description": "車端回報時刻"
          },
          "observed_clock": {
            "type": "string",
            "example": "08:59:58"
          },
          "data_age_seconds": {
            "type": "number",
            "example": 2,
            "description": "generated_at 減 observed_at。失聯是以「資料變舊」呈現而非欄位消失，不可只用欄位是否存在判斷"
          }
        },
        "required": [
          "vehicle_code",
          "order_id",
          "trip_code",
          "route_code",
          "route_name",
          "vehicle_phase",
          "arrival_state",
          "eta_seconds",
          "eta_at",
          "eta_clock",
          "distance_to_station_m",
          "plan",
          "observed_at",
          "observed_clock",
          "data_age_seconds"
        ]
      },
      "StationEtaGroupDto": {
        "type": "object",
        "properties": {
          "station_id": {
            "type": "string",
            "example": "station_4"
          },
          "station_name": {
            "type": "string",
            "example": "T3上行"
          },
          "etas": {
            "description": "依 eta_at 由近到遠排序；當下無車駛近時為空陣列",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StationEtaEntryDto"
            }
          }
        },
        "required": [
          "station_id",
          "station_name",
          "etas"
        ]
      },
...
      "TimetableMetaDto": {
        "type": "object",
        "properties": {
          "shift_id": {
            "type": "string",
            "example": "OS-DRAFT-MSEEXIN9",
            "description": "這份班表的識別碼"
          },
          "name": {
            "type": "string",
            "example": "模擬正線"
          },
          "publish_status": {
            "type": "string",
            "enum": [
              "draft",
              "published"
            ],
            "description": "班表本身的發布狀態"
          },
          "source": {
            "type": "string",
            "enum": [
              "published",
              "draft_fallback"
            ],
            "description": "published＝讀到已發布班表；draft_fallback＝庫內沒有已發布的，退而取用最新草稿，此時時刻僅供參考，不得作為正式時刻使用"
          },
          "generated_at": {
            "type": "object",
            "nullable": true,
            "example": "2026-08-19T00:53:59.220Z",
            "description": "班表產生時刻"
          },
          "updated_at": {
            "type": "string",
            "example": "1787115236800",
            "description": "最後更新（Epoch 毫秒字串）"
          },
          "map_id": {
            "type": "object",
            "nullable": true,
            "description": "這份班表所依據的地圖"
          }
        },
        "required": [
          "shift_id",
          "name",
          "publish_status",
          "source",
          "generated_at",
          "updated_at",
          "map_id"
        ]
      },
      "TimetableFilterDto": {
        "type": "object",
        "properties": {
          "from": {
            "type": "string",
            "example": "00:00:00"
          },
          "to": {
            "type": "string",
            "example": "24:00:00"
          },
          "from_second": {
            "type": "number",
            "example": 0,
            "description": "當日第幾秒"
          },
          "to_second": {
            "type": "number",
            "example": 86400
          }
        },
        "required": [
          "from",
          "to",
          "from_second",
          "to_second"
        ]
      },
      "TimetableTripStationDto": {
        "type": "object",
        "properties": {
          "order": {
            "type": "number",
            "example": 1,
            "description": "站序，自 1 起算"
          },
          "station_id": {
            "type": "string",
            "example": "station_4"
          },
          "station_name": {
            "type": "string",
            "example": "T3上行"
          },
          "role": {
            "type": "string",
            "enum": [
              "origin",
              "intermediate",
              "terminal"
            ],
            "description": "origin＝只看 departure（arrival 固定 null，抵達屬上一卡末站）；intermediate＝arrival ＋ departure；terminal＝arrival ＋ dwell_complete（完成靠站，不稱出發）"
          },
          "arrival": {
            "type": "object",
            "nullable": true,
            "example": "00:10:40",
            "description": "計畫抵達；起站為 null"
          },
          "departure": {
            "type": "object",
            "nullable": true,
            "example": "00:07:40",
            "description": "計畫離站；末站為 null"
          },
          "dwell_complete": {
            "type": "object",
            "nullable": true,
            "example": "00:11:30",
            "description": "靠站完成（＝該班次卡結束）。僅末站有值"
          },
          "base_dwell_seconds": {
            "type": "number",
            "example": 40,
            "description": "設定的靠站秒數，不含緩衝"
          },
          "dwell_seconds": {
            "type": "number",
            "example": 50,
            "description": "有效靠站秒數，含緩衝"
          },
          "travel_to_next_seconds": {
            "type": "object",
            "nullable": true,
            "example": 100,
            "description": "到下一站的行駛秒數；末站為 null"
          }
        },
        "required": [
          "order",
          "station_id",
          "station_name",
          "role",
          "arrival",
          "departure",
          "dwell_complete",
          "base_dwell_seconds",
          "dwell_seconds",
          "travel_to_next_seconds"
        ]
      },
      "TimetableTripDto": {
        "type": "object",
        "properties": {
          "trip_code": {
            "type": "string",
            "example": "ST0007",
            "description": "路線代號＋發車時刻 HHMM"
          },
          "block_id": {
            "type": "string",
            "example": "task-1786018029893-faqq",
            "description": "班次卡的內部識別碼"
          },
          "timeline_row": {
            "type": "number",
            "example": 1,
            "description": "這一趟屬於班表上的第幾列"
          },
          "task_type": {
            "type": "string",
            "example": "passenger",
            "description": "回傳全部任務類型：passenger 正線、servicing 保養、inspection 行檢、charging 充電、standby 待命、dispatch 調度等。整備類通常無站序，stations 為空陣列"
          },
          "label": {
            "type": "string",
            "example": "正線"
          },
          "source": {
            "type": "string",
            "example": "template_bar",
            "description": "這張卡的產生來源"
          },
          "route_id": {
            "type": "object",
            "nullable": true
          },
          "route_code": {
            "type": "object",
            "nullable": true,
            "example": "ST"
          },
          "route_name": {
            "type": "object",
            "nullable": true,
            "example": "S2W上行 > T3上行"
          },
          "card_start": {
            "type": "string",
            "example": "00:07:40",
            "description": "班次卡開始（＝起站 departure）"
          },
          "card_end": {
            "type": "string",
            "example": "00:11:30",
            "description": "班次卡結束（＝末站 dwell_complete）"
          },
          "card_start_second": {
            "type": "number",
            "example": 460
          },
          "card_end_second": {
            "type": "number",
            "example": 690
          },
          "stations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TimetableTripStationDto"
            }
          }
        },
        "required": [
          "trip_code",
          "block_id",
          "timeline_row",
          "task_type",
          "label",
          "source",
          "route_id",
          "route_code",
          "route_name",
          "card_start",
          "card_end",
          "card_start_second",
          "card_end_second",
          "stations"
        ]
      },
      "TimetableTripsResponseDto": {
        "type": "object",
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/TimetableMetaDto"
          },
          "filter": {
            "$ref": "#/components/schemas/TimetableFilterDto"
          },
          "trip_count": {
            "type": "number",
            "example": 26
          },
          "trips": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TimetableTripDto"
            }
          }
        },
        "required": [
          "meta",
          "filter",
          "trip_count",
          "trips"
        ]
      },
      "StationEtaFilterDto": {
        "type": "object",
        "properties": {
          "from": {
            "type": "string",
            "example": "00:00:00"
          },
          "to": {
            "type": "string",
            "example": "24:00:00"
          },
          "from_second": {
            "type": "number",
            "example": 0,
            "description": "當日第幾秒"
          },
          "to_second": {
            "type": "number",
            "example": 86400
          },
          "station_id": {
            "type": "object",
            "nullable": true,
            "example": "station_4"
          },
          "passenger_stops_only": {
            "type": "boolean",
            "example": true,
            "description": "固定為 true：只回正線班次的站點事件"
          }
        },
        "required": [
          "from",
          "to",
          "from_second",
          "to_second",
          "station_id",
          "passenger_stops_only"
        ]
      },
      "StationEtaEventDto": {
        "type": "object",
        "properties": {
          "station_id": {
            "type": "string",
            "example": "station_4"
          },
          "station_alias": {
            "type": "string",
            "example": "T3上行",
            "description": "地圖上的停靠點別名，顯示用主標題"
          },
          "station_name": {
            "type": "string",
            "example": "T3上行",
            "description": "路線站序上的名稱"
          },
          "role": {
            "type": "string",
            "enum": [
              "origin",
              "intermediate",
              "terminal"
            ]
          },
          "eta_arrive": {
            "type": "object",
            "nullable": true,
            "example": "00:10:40",
            "description": "計畫抵達；起站為 null"
          },
          "eta_depart": {
            "type": "object",
            "nullable": true,
            "example": "00:11:30",
            "description": "起站／途經＝出發；末站＝靠站完成（即該班次卡結束）"
          },
          "eta_arrive_second": {
            "type": "object",
            "nullable": true,
            "example": 640
          },
          "eta_depart_second": {
            "type": "object",
            "nullable": true,
            "example": 690
          },
          "non_stop": {
            "type": "boolean",
            "example": false,
            "description": "到站與離站同一秒（不停靠／途經）"
          },
          "base_dwell_seconds": {
            "type": "number",
            "example": 36
          },
          "dwell_seconds": {
            "type": "number",
            "example": 50
          },
          "buffer_seconds": {
            "type": "number",
            "example": 14,
            "description": "有效停靠 − 設定停靠，恆 ≥ 0"
          },
          "trip_code": {
            "type": "string",
            "example": "ST0007"
          },
          "block_id": {
            "type": "string",
            "example": "b-st"
          },
          "timeline_row": {
            "type": "number",
            "example": 1
          },
          "route_code": {
            "type": "object",
            "nullable": true,
            "example": "ST"
          },
          "route_name": {
            "type": "object",
            "nullable": true,
            "example": "S2W上行 > T3上行"
          },
          "vehicle_id": {
            "type": "string",
            "nullable": true,
            "description": "<strong>固定為 null。</strong>計畫值不含車輛對應——哪一列由哪一台車擔任，是當日才決定的。要車輛識別請改用即時 ETA 介面。"
          }
        },
        "required": [
          "station_id",
          "station_alias",
          "station_name",
          "role",
          "eta_arrive",
          "eta_depart",
          "eta_arrive_second",
          "eta_depart_second",
          "non_stop",
          "base_dwell_seconds",
          "dwell_seconds",
          "buffer_seconds",
          "trip_code",
          "block_id",
          "timeline_row",
          "route_code",
          "route_name",
          "vehicle_id"
        ]
      },
      "StationEtasResponseDto": {
        "type": "object",
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/TimetableMetaDto"
          },
          "filter": {
            "$ref": "#/components/schemas/StationEtaFilterDto"
          },
          "eta_count": {
            "type": "number",
            "example": 4
          },
          "etas": {
            "description": "依時間排序的站點事件流",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StationEtaEventDto"
            }
          },
          "stations": {
            "description": "同一份事件依停靠點分組。<strong>會列出地圖上全部停靠點</strong>，包含目前沒有任何班次的（etas 為空陣列），方便站端固定顯示版面而不必自己補齊清單。",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StationEtaGroupDto"
            }
          }
        },
        "required": [
          "meta",
          "filter",
          "eta_count",
          "etas",
          "stations"
        ]
      },
...
    }
  }
}